pyproject.toml
[tool.poetry]
name = "qq-github-bot"
version = "1.0.5"
description = "GitHub Bot for QQ"
authors = ["yanyongyu <yyy@yyydl.top>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
psutil = "^5.7.2"
Jinja2 = "^3.0.0"
aerich = "^0.6.3"
unidiff = "^0.7.0"
humanize = "^4.0.0"
sentry-sdk = "^1.0.0"
playwright = "^1.17.2"
mdit-py-emoji = "^0.1.0"
typing-extensions = "^4.0.0"
nonebot-adapter-github = "^0.3.0"
nonebot-adapter-onebot = "^2.1.4"
redis = { version = "^4.0.0", extras = ["hiredis"] }
tortoise-orm = { version = "^0.19.2", extras = ["asyncpg"] }
nonebot2 = { version = "^2.0.0", extras = ["httpx", "fastapi"] }
markdown-it-py = { version = "^3.0.0", extras = ["linkify", "plugins"] }

[tool.poetry.group.dev.dependencies]
isort = "^5.9.3"
black = "^23.1.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"
ruff = ">=0.0.291,<0.1.0"

[tool.poetry.group.deploy.dependencies]
gunicorn = "^20.1.0"
uvicorn = ">=0.18.3,<1.0.0"

[tool.black]
preview = true
line-length = 88
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = "typing_extensions"

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

line-length = 88
target-version = "py38"

[tool.ruff.extend-per-file-ignores]
"./scripts/*" = ["T201"]
"./docker/*" = ["T201"]

[tool.pyright]
pythonPlatform = "All"
executionEnvironments = [
  { root = "./src/plugins/nonebot_plugin_sentry", pythonVersion = "3.8" },
  { root = "./src/plugins/nonebot_plugin_status", pythonVersion = "3.8" },
  { root = "./", pythonVersion = "3.11" },
]

typeCheckingMode = "basic"
reportShadowedImports = false
disableBytesTypePromotions = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
